+++ assets/src/components/modals/ReminderModal.js
@@ -1,31 +1,11 @@
-import { connect } from "react-redux"
-import styled from "styled-components"
 import { withRouter } from "react-router-dom"
 import { compose } from "redux"
 
 import { Headline, Subline } from "../Text"
 import { Button } from "../Button"
+import { ModalContainer, ModalContent } from "./Modal"
 
 import success from "../../static/icons/success.svg"
-
-const ModalContainer = styled.div.attrs({
-  className: "flex items-center justify-center blue",
-})`
-  background-color: rgba(51, 51, 51, 0.5);
-  position: fixed;
-  width: 100vw;
-  height: 100vh;
-  z-index: 9999999;
-`
-
-const ModalContent = styled.div.attrs({
-  className: "bg-near-white pa7-ns pv7-ns ph5 pv6 tc",
-})`
-  width: 90%;
-  @media screen and (min-width: 30em) {
-    width: 35vw;
-  }
-`
 
 const ReminderModal = ({ history }) => (
   <ModalContainer onClick={() => history.push("/login")}>
